Created a simulation class and rounded up parameters
Created a simulation class of which will handel the disease, the population, and interactions between the two.

Also created and moved all attribute parameters used into an attribute class to be imported. This will make future changes in parameters easier to make. I also chose to use a class in case we want our infection formula to be inheriented and such.